+++ pyevr/__init__.py
@@ -1,5 +1,5 @@
-# -*- coding: utf-8 -*-
-
 """Top-level package for pyevr."""
 
 from .client import EVRClient
+
+__all__ = ["EVRClient"]